On branch beta_from_Taylor_coeffs  Changes to be committed: 	modified:   pyLaserPulse/abstract_bases.py 	modified:   pyLaserPulse/utils.py utils: created new function which returns a Taylor expansion given coefficients, an axis, and the centre of the axis.
abstract_bases: Added new method to fibre_base which allows the
user to override the dispersion curve calculated from the fibre
geometry with Taylor coefficients. This is useful when, for
example, something about the fibre geometry means that the
Saitoh and Gloge methods can't be used to estimate the fibre
dispersion accurately (e.g., if the core material is doped in
such a way that the material dispersion is sufficiently different
to that given by the Sellmeier equation).