On branch oscillator_testing  Your branch is up-to-date with 'origin/oscillator_testing'.
 Changes to be committed:
	modified:   examples/fibre_amplifiers/Yb_fibre_CPA_system.py
	modified:   examples/fibre_lasers/Yb_mode_locked_oscillator.py
	modified:   pyLaserPulse/base_components.py

Rabbit hole -- Noticed that the pulse ASE_scaling might not be being
recalculated when the repetition rate is changed. This does not seem
to be the case for the pump light, and will make a negligible
difference to the simulation results, but it will be good to make sure
it is correct (and to correct it if it is not).

To do for oscillator development (once the above is complete):
1) Pump and ASE members of the gain fibre classes must be reset
   with each round trip. (pump objects, stacks, any other container
   or structure that is re-used between round trips).
2) Repeat 1, but include the boundary value solver. Everything must
   be reset for each round trip -- assumption that ASE in pump
   fields is negligible. Not the case, but the alternative is out
   of scope at the moment.
3) Consider mitigating the assumption in 2. This will involve a
   substantial amount of work.