On branch oscillator_testing  Your branch is up-to-date with 'origin/oscillator_testing'.
 Changes to be committed:
	modified:   examples/fibre_lasers/Yb_mode_locked_oscillator.py
	modified:   pyLaserPulse/abstract_bases.py
	modified:   pyLaserPulse/optical_assemblies.py

Co-/counter-propagating gain in a mode-locked oscillator simulation
now working (although no example with mode-locked solutions has been
found yet). Still not thoroughly tested, but the gain appears to
be working well.

The co_ASE flag for each amplifier needs testing. I am confident that
the implementation is correct, but I need to check that the
transmission window of each component is being applied to the pump/ASE
channels adequately.

The one component which does NOT have this functionality at the moment
is the new saturable Bragg reflector class in base_components.py. This
is the next thing to work on.