On branch pcf_gain_debug  Your branch is up to date with 'origin/pcf_gain_debug'.
 Changes to be committed:
	modified:   pyLaserPulse/catalogue_components/active_fibres.py

Reasonable performance from the NKT DC 200 40 PZ YB model.

This commit solves the following issue:
The usual PCF model is not too accurate for this fibre type. Although
the dispersion profile looks very reasonable and the pump overlap with
the core is accurate when the nominal core diameter of 40 um is used, the
mode field diameter (and, subsequently, overlap with the core) of the signal
light does not match the published specifications.

This is solved using the following workaround:
1) Calculate all signal parameters using a 20 um core diameter.
2) Reset the core diameter (and radius) to 40 um if cladding pumping.
3) Recalculate the cladding light overlap parameters for the new core
   diameter.

Adjusting the doping concentration of the core then gives:
10 - 11 dB/m (1 mW 976 nm through 1 m of fibre, boundary value solver,
cladding guided, no signal), matching published specification, and a
~6% overlap of the cladding light with the doped core, which matches
approximate calculations using the ratios of the core and cladding areas.