On branch oscillator_testing  Changes to be committed: 	modified:   pyLaserPulse/catalogue_components/passive_fibres.py
Added Coherent PLMA-GDF-25/400-M to passive_fibres.py

Same definition as PLMA-GDF-25/250-M, but having this new class for
a passive fibre with the same core parameters but different cladding
parameters will make models easier to read even if the performance
is identical. Illustration:

Model 1: GDF 25/250 spliced to YDF 25/400
Model 2: GDF 25/400 spliced to YDF 25/400

The performance of both will be identical in pyLaserPulse, but it
is unlikely that someone would choose to configure their experiment
as per Model 1.